[IPVS] Add AF to Service structure and expose value for Fwmark
Linux kernel/IPVS can apply fwmakr to only one version of IP at a time.
Previous code was already supporting setting the version
but was not exposing it on the read path.
This diff fix that and update various test that fail due to change in structure

I also tested runnning examples/ipvs_dump on a server using FWMark.